backend/requirements.txt
# Core Framework
django>=5.0.0
django-ninja>=1.1.0
django-ninja-extra>=0.20.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Agent & LLM
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0

# Text-to-SQL
vanna>=0.3.4
chromadb>=0.4.22

# Web Search
tavily-python>=0.3.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# CORS
django-cors-headers>=4.3.0

# Static files
whitenoise>=6.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
faker>=22.0.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
